enable to dynamic routing by pageId
