+++ package.json
@@ -44,6 +44,7 @@
     "gatsby-background-image-es5": "^1.1.1",
     "gatsby-image": "^2.2.1",
     "gatsby-plugin-canonical-urls": "^2.1.0",
+    "gatsby-plugin-create-client-paths": "^2.3.10",
     "gatsby-plugin-emotion": "^4.3.4",
     "gatsby-plugin-firebase": "^0.2.0-beta.4",
     "gatsby-plugin-manifest": "^2.4.21",
