+++ src/pages/register.tsx
@@ -29,11 +29,10 @@ const invitationSignIn = async (invitationCode: string, login: (loginUser: Login
   // FIXME move to cloud functions
   const invitation = await AccountService.acceptInvitation(invitationCode)
   await CommunityService.join(invitation.hostCommunity, user.uid)
-  await PersonService.addPage(user.uid, user.name, user.img)
-  login(LoginUserModel.fromUserCredential(user))
-  // TODO:
-  // navigate(`/persons/${loginUser.uid}`)
-  navigate('/persons/baiUK5z4NYhFscfuwpJuT7NTwZs2')
+  await PersonService.addPage(user.uid, user.uid, user.name, user.img)
+  const loginUser = LoginUserModel.fromUserCredential(user)
+  login(loginUser)
+  navigate(`/persons/${loginUser.uid}`)
 }
 const InvitationCode = ({ invitation, setInvitation }) => {
   return (
